+++ HISTORY.md
@@ -7,6 +7,9 @@
 * **Documentation Updates:**
 * **Translation Updates:**
 
+### 2022-03-09
+* new dev version
+
 ## 7.3.0:
 * **New Features:**
     * variadic associative infix reporters