+++ HISTORY.md
@@ -5,6 +5,7 @@
     * export / import sprite-local custom block definitions, under construction
 * **Notable Changes:**
 * **Notable Fixes:**
+    * guard against broken SVG costumes when loading a project
 * **Documentation Updates:**
 * **Translation Updates:**
 
@@ -13,6 +14,7 @@
 * byob, store, gui: export / import sprite-local custom block definitions, under construction
 * byob, gui: adapted library import dialog to the new format
 * byob: added collecting dependencies for sprite-local custom blocks
+* objects: guard against broken SVG costumes when loading a project
 
 ## 7.3.0:
 * **New Features:**